Remove English inflections (fixes #13)
Because of how much the rest of Rails relies on the Inflector,
overriding the English rules was a futile effort and it feels like it's
time to give up. Rails uses these rules for everything from deriving
table names to model names to controller names, including generating new
ones. This leads to usually incorrect results as it contains many rules
to prevent double pluralization or double singularization. Inflections
will, instead, act as a repository for rules in other languages.

Signed-off-by: David Celis <me@davidcel.is>